fix: args in long (ex int)

+++ models/args.h
@@ -6,7 +6,7 @@
 /*   By: phenry <phenry@student.42mulhouse.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/07/05 19:47:10 by phenry            #+#    #+#             */
-/*   Updated: 2026/07/18 03:06:12 by phenry           ###   ########.fr       */
+/*   Updated: 2026/07/18 12:03:26 by phenry           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,13 +17,13 @@
 
 typedef struct s_args
 {
-	int			number_of_coders;
-	int			time_to_burnout;
-	int			time_to_compile;
-	int			time_to_debug;
-	int			time_to_refactor;
-	int			number_of_compiles_required;
-	int			dongle_cooldown;
+	long		number_of_coders;
+	long		time_to_burnout;
+	long		time_to_compile;
+	long		time_to_debug;
+	long		time_to_refactor;
+	long		number_of_compiles_required;
+	long		dongle_cooldown;
 	t_strategy	strategy;
 }	t_args;
